requirements.txt
sqlalchemy=0.9.9
PyYaml
simple-crypt
colorama
requests
colorlog
PyMySQL
prettytable
inlinestyler
xmldataset
